Move checkpointer to agents/memory folder
